(2/3) Make TimeKeep available to keep time offset
TimeKeep is utility to keep/restore RTC offset for Qualcomm device, where
RTC is not setable. It works by keeping the offset between the RTC time and
system time in a property (or a file) and restore it at boot.

This commit add Sony's vendor subdirectory to the subdirs list. As TimeKeep
is originaly from Sony's AOSP, it's kept in Sony's vendor directory.

Change-Id: I6673e7447f2715aabb8926bb39f491d42f848145